+++ trading-api/src/routes.rs
@@ -4,13 +4,12 @@ use axum::{
     Json,
 };
 use serde_json::json;
-use std::sync::Arc;
 use trading_common::{
     error::AppError,
     models::{BuyRequest, BuyResponse, SellRequest, SellResponse},
-    process_buy_request, process_sell_request,
+    pumpdotfun::{buy::process_buy_request, sell::process_sell_request},
     utils::get_server_keypair,
-    CopyTradeSettings, SupabaseClient, TrackedWallet, TransactionLog,
+    CopyTradeSettings, TrackedWallet, TransactionLog,
 };
 use uuid::Uuid;
