+++ trading-common/src/lib.rs
@@ -2,9 +2,9 @@ pub mod constants;
 pub mod database;
 pub mod error;
 pub mod models;
-pub mod pump_fun;
+pub mod pumpdotfun; // New module
 pub mod utils;
 pub use constants::*;
 pub use database::SupabaseClient;
 pub use models::{ClientTxInfo, CopyTradeSettings, TrackedWallet, TransactionLog, TransactionType};
-pub use pump_fun::*;
+pub use pumpdotfun::*;
