+++ trading-common/src/pumpdotfun/constants.rs
@@ -1,6 +1,13 @@
+use solana_sdk::pubkey::Pubkey;
+
+pub const UNIT_PRICE: u64 = 1_000_000; // Match Python exactly
+pub const UNIT_BUDGET: u32 = 100_000; // Match Python exactly
+pub const LAMPORTS_PER_SOL: u64 = 1_000_000_000;
 pub const BONDING_CURVE_MARGIN_OF_ERROR: f64 = 0.01; // 1%
 pub const BUY: u8 = 0;
 pub const SELL: u8 = 1;
+
+// Program IDs and account addresses
 pub const PUMP_FUN_PROGRAM_ID: Pubkey =
     solana_sdk::pubkey!("6EF8rrecthR5Dkzon8Nwu78hRvfCKubJ14M5uBEwF6P");
 pub const GLOBAL: Pubkey = solana_sdk::pubkey!("4wTV1YmiEkRvAtNtsSGPtUrqRYQMe5SKy2uB4Jjaxnjf");
@@ -14,5 +21,6 @@ pub const EVENT_AUTHORITY: Pubkey =
 pub const ASSOCIATED_TOKEN_PROGRAM_ID: Pubkey =
     solana_sdk::pubkey!("ATokenGPvbdGVxr1b2hvZbsiqW5xWH25efTNsLJA8knL");
 
+// Instruction discriminators
 pub const BUY_DISCRIMINATOR: [u8; 8] = [102, 6, 61, 18, 1, 218, 235, 234]; // "66063d1201daebea"
 pub const SELL_DISCRIMINATOR: [u8; 8] = [51, 230, 133, 164, 1, 127, 131, 173]; // "33e685a4017f83ad"
